+++ src/app/cli/src/cli_parser.rs
@@ -1153,6 +1153,10 @@ pub fn cli() -> Command {
                                     .long("http-port")
                                     .value_parser(value_parser!(u16))
                                     .help("Expose HTTP+GraphQL server on specific port"),
+                                Arg::new("get-token")
+                                    .long("get-token")
+                                    .action(ArgAction::SetTrue)
+                                    .help("Output a JWT token you can use to authorize API queries"),
                             ])
                             .after_help(indoc::indoc!(
                                 r#"