+++ src/app/cli/src/commands/system_api_server_run_command.rs
@@ -12,14 +12,17 @@ use std::sync::Arc;
 
 use console::style as s;
 use dill::Catalog;
-use kamu::ENV_VAR_KAMU_JWT_SECRET;
+use internal_error::ResultIntoInternal;
+use kamu::domain::auth::AuthenticationService;
+use kamu::domain::CurrentAccountSubject;
+use kamu::{set_random_jwt_secret, ENV_VAR_KAMU_JWT_SECRET};
 use kamu_adapter_oauth::{
     ENV_VAR_KAMU_AUTH_GITHUB_CLIENT_ID,
     ENV_VAR_KAMU_AUTH_GITHUB_CLIENT_SECRET,
 };
 
 use super::{CLIError, Command};
-use crate::{check_env_var_set, OutputConfig};
+use crate::{accounts, check_env_var_set, OutputConfig};
 
 pub struct APIServerRunCommand {
     catalog: Catalog,
@@ -27,6 +30,8 @@ pub struct APIServerRunCommand {
     output_config: Arc<OutputConfig>,
     address: Option<IpAddr>,
     port: Option<u16>,
+    get_token: bool,
+    account_subject: Arc<CurrentAccountSubject>,
 }
 
 impl APIServerRunCommand {
@@ -36,6 +41,8 @@ impl APIServerRunCommand {
         output_config: Arc<OutputConfig>,
         address: Option<IpAddr>,
         port: Option<u16>,
+        get_token: bool,
+        account_subject: Arc<CurrentAccountSubject>,
     ) -> Self {
         Self {
             catalog,
@@ -43,11 +50,16 @@ impl APIServerRunCommand {
             output_config,
             address,
             port,
+            get_token,
+            account_subject,
         }
     }
 
     fn check_required_env_vars(&self) -> Result<(), CLIError> {
-        check_env_var_set(ENV_VAR_KAMU_JWT_SECRET)?;
+        match check_env_var_set(ENV_VAR_KAMU_JWT_SECRET) {
+            Ok(_) => {}
+            Err(_) => set_random_jwt_secret(),
+        };
 
         if self.multi_tenant_workspace {
             check_env_var_set(ENV_VAR_KAMU_AUTH_GITHUB_CLIENT_ID)?;
@@ -56,6 +68,33 @@ impl APIServerRunCommand {
 
         Ok(())
     }
+
+    async fn get_access_token(&self) -> Result<String, CLIError> {
+        let current_account_name = match self.account_subject.as_ref() {
+            CurrentAccountSubject::Logged(l) => l.account_name.clone(),
+            CurrentAccountSubject::Anonymous(_) => {
+                unreachable!("Cannot launch API server with anonymous account")
+            }
+        };
+        let login_credentials = accounts::PasswordLoginCredentials {
+            login: current_account_name.to_string(),
+            // Note: note a mistake, use identical login and password, equal to account name
+            password: current_account_name.to_string(),
+        };
+
+        let auth_svc = self.catalog.get_one::<dyn AuthenticationService>().unwrap();
+        let access_token = auth_svc
+            .login(
+                &accounts::LOGIN_METHOD_PASSWORD.to_string(),
+                serde_json::to_string::<accounts::PasswordLoginCredentials>(&login_credentials)
+                    .int_err()?,
+            )
+            .await
+            .int_err()?
+            .access_token;
+
+        Ok(access_token)
+    }
 }
 
 #[async_trait::async_trait(?Send)]
@@ -63,6 +102,7 @@ impl Command for APIServerRunCommand {
     async fn run(&mut self) -> Result<(), CLIError> {
         // Check required env variables are present before starting API server
         self.check_required_env_vars()?;
+        let access_token = self.get_access_token().await?;
 
         // TODO: Cloning catalog is too expensive currently
         let api_server = crate::explore::APIServer::new(
@@ -87,6 +127,14 @@ impl Command for APIServerRunCommand {
                 s(format!("http://{}", api_server.local_addr())).bold(),
             );
             eprintln!("{}", s("Use Ctrl+C to stop the server").yellow());
+
+            if self.get_token {
+                eprintln!(
+                    "{} {}",
+                    s("JWT token:").green().bold(),
+                    s(access_token).dim()
+                );
+            }
         }
 
         api_server.run().await.map_err(|e| CLIError::critical(e))?;